Add: Moved new files to a new folder
