+++ src/main/java/se/branko/dependencyInversion/Discount.java
@@ -1,4 +1,4 @@
-package se.branko;
+package se.branko.dependencyInversion;
 
 public interface Discount {
     double applyDiscount(double total);
