+++ src/main/java/se/branko/dependencyInversion/PercentageDiscount.java
@@ -1,4 +1,4 @@
-package se.branko;
+package se.branko.dependencyInversion;
 
 public class PercentageDiscount implements Discount{
