+++ src/main/java/se/branko/dependencyInversion/ShoppingCart.java
@@ -1,4 +1,4 @@
-package se.branko;
+package se.branko.dependencyInversion;
 
 public class ShoppingCart {
     private Discount discount;
